+++ bin/client/src/witgen.rs
@@ -16,9 +16,11 @@ use crate::witness::{BlobWitnessProvider, OracleWitnessProvider};
 use alloy_primitives::{Address, B256};
 use kailua_build::KAILUA_FPVM_ID;
 use kailua_common::blobs::BlobWitnessData;
+use kailua_common::boot::StitchedBootInfo;
 use kailua_common::executor::Execution;
 use kailua_common::journal::ProofJournal;
-use kailua_common::witness::{StitchedBootInfo, Witness, WitnessOracle};
+use kailua_common::oracle::WitnessOracle;
+use kailua_common::witness::Witness;
 use kona_derive::prelude::BlobProvider;
 use kona_preimage::CommsClient;
 use kona_proof::FlushableCache;
