+++ bin/host/src/prove.rs
@@ -22,9 +22,9 @@ use async_channel::Sender;
 use kailua_build::KAILUA_FPVM_ID;
 use kailua_client::proof::{proof_file_name, read_proof_file};
 use kailua_client::proving::ProvingError;
+use kailua_common::boot::StitchedBootInfo;
 use kailua_common::client::stitching::{split_executions, stitch_boot_info};
 use kailua_common::executor::{exec_precondition_hash, Execution};
-use kailua_common::witness::StitchedBootInfo;
 use kona_genesis::RollupConfig;
 use kona_proof::BootInfo;
 use risc0_zkvm::Receipt;
