+++ bin/host/src/tasks.rs
@@ -19,8 +19,8 @@ use alloy_primitives::B256;
 use anyhow::Context;
 use async_channel::{Receiver, Sender};
 use kailua_client::proving::ProvingError;
+use kailua_common::boot::StitchedBootInfo;
 use kailua_common::executor::Execution;
-use kailua_common::witness::StitchedBootInfo;
 use kona_genesis::RollupConfig;
 use risc0_zkvm::Receipt;
 use std::cmp::Ordering;
