+++ crates/common/src/oracle/mod.rs
@@ -17,8 +17,10 @@ pub mod vec;
 
 use alloy_primitives::keccak256;
 use kona_preimage::errors::{PreimageOracleError, PreimageOracleResult};
-use kona_preimage::{PreimageKey, PreimageKeyType};
+use kona_preimage::{CommsClient, PreimageKey, PreimageKeyType};
+use kona_proof::FlushableCache;
 use risc0_zkvm::sha::{Impl as SHA2, Sha256};
+use std::fmt::Debug;
 
 /// Determines if a given `PreimageKeyType` requires validation.
 ///
@@ -95,4 +97,32 @@ pub fn validate_preimage(key: &PreimageKey, value: &[u8]) -> PreimageOracleResul
         }
     }
     Ok(())
+}
+
+/// A trait representing a Witness Oracle which manages and validates cryptographic preimages.
+///
+/// The `WitnessOracle` trait provides functionality to interact with and manage preimages.
+/// Preimages are key-value pairs where the key is typically an identifier for the data,
+/// and the value is the data itself stored as a `Vec<u8>`. This trait incorporates several features
+/// including validation, insertion, count retrieval, and finalization of the stored preimages.
+///
+/// # Provided Methods
+/// - `preimage_count`: Retrieve the number of preimages stored.
+/// - `validate_preimages`: Validates that all stored preimages satisfy predefined constraints.
+/// - `insert_preimage`: Inserts a new preimage into the oracle.
+/// - `finalize_preimages`: Prepares the oracle's preimages for use under defined shard size and validation settings.
+pub trait WitnessOracle: CommsClient + FlushableCache + Send + Sync + Debug + Default {
+    /// Returns the count of preimages stored in the oracle.
+    fn preimage_count(&self) -> usize;
+
+    /// Ensures that the preimages stored in the oracle meet the required criteria or constraints
+    /// defined by each `PreimageKeyType`. If the validation fails, an error is returned.
+    fn validate_preimages(&self) -> anyhow::Result<()>;
+
+    /// Inserts a preimage into the oracle.
+    fn insert_preimage(&mut self, key: PreimageKey, value: Vec<u8>);
+
+    /// This method finalizes the process of preparing the oracle preimages for a specific shard
+    /// size and optional validation cache.
+    fn finalize_preimages(&mut self, shard_size: usize, with_validation_cache: bool);
 }
